Basic testing for score guides.
More work
