+++ ash/Makefile
@@ -1,0 +1,3 @@
+install-deps:
+	pip install -r requirements.txt
+	python setup.py develop
