+++ ash/ash.egg-info/PKG-INFO
@@ -1,0 +1,11 @@
+Metadata-Version: 1.0
+Name: ash
+Version: 1.0
+Summary: Shared code for (FL|D)ash.
+Home-page: UNKNOWN
+Author: David Dynerman
+Author-email: david.dynerman@czbiohub.org
+License: UNKNOWN
+Description-Content-Type: UNKNOWN
+Description: UNKNOWN
+Platform: UNKNOWN
