+++ ash/setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup
+
+setup(
+    name='ash',
+    version='1.0',
+    description='Shared code for (FL|D)ash.',
+    author='David Dynerman',
+    author_email='david.dynerman@czbiohub.org',
+    packages=['ash'],
+)
