+++ dashit/Makefile
@@ -1,0 +1,2 @@
+install:
+	cd dashit-seq && $(MAKE) install-develop
