+++ dashit/dashit-seq/MANIFEST
@@ -1,0 +1,4 @@
+# file GENERATED by distutils, do NOT edit
+setup.py
+dashit-seq/__init__.py
+dashit-seq/dashit-seq.py
