+++ dashit/dashit-seq/Makefile
@@ -1,0 +1,3 @@
+install-develop:
+	pip install -r requirements.txt
+	python setup.py develop
