+++ dashit/dashit-seq/dashit_seq.egg-info/PKG-INFO
@@ -1,0 +1,11 @@
+Metadata-Version: 1.0
+Name: dashit-seq
+Version: 1.0
+Summary: DASHit Seq.
+Home-page: UNKNOWN
+Author: David Dynerman
+Author-email: david.dynerman@czbiohub.org
+License: UNKNOWN
+Description-Content-Type: UNKNOWN
+Description: UNKNOWN
+Platform: UNKNOWN
