+++ dashit/dashit-seq/requirements.txt
@@ -1,0 +1,54 @@
+ansicolors==1.1.8
+appnope==0.1.0
+beautifulsoup4==4.6.0
+biopython==1.70
+bleach==2.1.1
+bs4==0.0.1
+certifi==2017.11.5
+chardet==3.0.4
+decorator==4.1.2
+entrypoints==0.2.3
+html5lib==1.0b10
+idna==2.6
+ipykernel==4.6.1
+ipython==6.2.1
+ipython-genutils==0.2.0
+ipywidgets==7.0.4
+jedi==0.11.0
+Jinja2==2.10
+jsonschema==2.6.0
+jupyter==1.0.0
+jupyter-client==5.1.0
+jupyter-console==5.2.0
+jupyter-core==4.4.0
+MarkupSafe==1.0
+mistune==0.8.1
+nbconvert==5.3.1
+nbformat==4.4.0
+notebook==5.2.1
+numpy==1.13.3
+ortools==6.5.4527
+pandocfilters==1.4.2
+parso==0.1.0
+pexpect==4.3.0
+pickleshare==0.7.4
+prompt-toolkit==1.0.15
+protobuf==3.3.0
+ptyprocess==0.5.2
+Pygments==2.2.0
+pysam==0.13
+python-dateutil==2.6.1
+pyzmq==16.0.3
+qtconsole==4.3.1
+requests==2.18.4
+scipy==1.0.0
+simplegeneric==0.8.1
+six==1.11.0
+terminado==0.6
+testpath==0.3.1
+tornado==4.5.2
+traitlets==4.3.2
+urllib3==1.22
+wcwidth==0.1.7
+webencodings==0.5.1
+widgetsnbextension==3.0.7
