+++ dashit/dashit-seq/setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup
+
+setup(
+    name='dashit-seq',
+    version='1.0',
+    description='DASHit Seq.',
+    author='David Dynerman',
+    author_email='david.dynerman@czbiohub.org',
+    packages=['dashit-seq'],
+)
