+++ misc/tests/data/test_expected_reads_dashed.fasta
@@ -1,0 +1,4 @@
+>HWI-700L170C:151:HAFY7ADXX:1:1101:1923:2223 1:N:0:ATGTCAGGGTAATC
+ACCCTGGGGTTAGTATAGCTTAGTTAAACTTTCGTTTATTGCTAAAGGTTAATCACTGCTGTTTCCCGTGGGGGTGTGGCTAGGCTAAGCGTTTTGAGCTGCATTGCTGCGTGCTTGATGCTTGTCCCTTTTGAT
+>HWI-700L170C:151:HAFY7ADXX:1:1101:2077:2218 1:N:0:ATGTCAGGGTAATC
+GGTTAATCGTGTGACCGCGGTGGCTGGCACGAAATTGACCAACCCTGGGGTTAGTATAGCTTAGTTAAACTTTCGTTTATTGCTAAAGGTTAATCACTGCTGTTTCCCGTGGGGGTGTGGCTAGGCTAAGCGTTT
