+++ misc/tests/data/test_expected_reads_undashed.fasta
@@ -1,0 +1,6 @@
+>HWI-700L170C:151:HAFY7ADXX:1:1101:1248:2121 1:N:0:ATGTCAGGGTAATC
+CTCCAGACCTGCCGGTTCCCATGGCTTCTACTTTCTCAACTTCAATGTCACATTCTCAAAAGGAAAAATTAAAGGATACTATCCATATAAGGAAAACAAATGGAAATAAAAATGCAGCTATTGACTTCTCTTTCA
+>HWI-700L170C:151:HAFY7ADXX:1:1101:1528:2086 1:N:0:ATGTCAGGGTAATC
+CTTAACAGAAGAGAGAATAAGGAGAAAGATTTACAATGAACTTGCCTTTTTCCTGTAAACCAGCCACATTCTTAAGCAAAACCTGAGGGTACATAAACATGAAAGGTTATTTCAAATAACCAGATGTTCATTTTA
+>HWI-700L170C:151:HAFY7ADXX:1:1101:2124:2175 1:N:0:ATGTCAGGGTAATC
+GTATAGGGGTTAGTCCTTGCTATATTATGCTTGGTTATAATTTTTCATCTTTCCCTTGCGGTACTATATCTATTGCGCCAGGTTTCAATTTCTAGAAATAACTTTGCAAGGAGAACCAAAGCTAAGACCCCCGAA
