+++ misc/tests/data/test_guides.csv
@@ -1,0 +1,5 @@
+Total number of reads: 16617695
+Site, Site index, Number of reads covered by site, cumulative number of reads covered
+AAACTTTCGTTTATTGCTAA, 87736, 1008161, 1008161
+AGGGTGAACTCACTGGAATG, 745060, 822244, 1830405
+TAGAGTCCATATCAACAATA, 3215257, 514966, 2345371
