+++ misc/tests/data/test_reads.fasta
@@ -1,0 +1,10 @@
+>HWI-700L170C:151:HAFY7ADXX:1:1101:1248:2121 1:N:0:ATGTCAGGGTAATC
+CTCCAGACCTGCCGGTTCCCATGGCTTCTACTTTCTCAACTTCAATGTCACATTCTCAAAAGGAAAAATTAAAGGATACTATCCATATAAGGAAAACAAATGGAAATAAAAATGCAGCTATTGACTTCTCTTTCA
+>HWI-700L170C:151:HAFY7ADXX:1:1101:1528:2086 1:N:0:ATGTCAGGGTAATC
+CTTAACAGAAGAGAGAATAAGGAGAAAGATTTACAATGAACTTGCCTTTTTCCTGTAAACCAGCCACATTCTTAAGCAAAACCTGAGGGTACATAAACATGAAAGGTTATTTCAAATAACCAGATGTTCATTTTA
+>HWI-700L170C:151:HAFY7ADXX:1:1101:1923:2223 1:N:0:ATGTCAGGGTAATC
+ACCCTGGGGTTAGTATAGCTTAGTTAAACTTTCGTTTATTGCTAAAGGTTAATCACTGCTGTTTCCCGTGGGGGTGTGGCTAGGCTAAGCGTTTTGAGCTGCATTGCTGCGTGCTTGATGCTTGTCCCTTTTGAT
+>HWI-700L170C:151:HAFY7ADXX:1:1101:2124:2175 1:N:0:ATGTCAGGGTAATC
+GTATAGGGGTTAGTCCTTGCTATATTATGCTTGGTTATAATTTTTCATCTTTCCCTTGCGGTACTATATCTATTGCGCCAGGTTTCAATTTCTAGAAATAACTTTGCAAGGAGAACCAAAGCTAAGACCCCCGAA
+>HWI-700L170C:151:HAFY7ADXX:1:1101:2077:2218 1:N:0:ATGTCAGGGTAATC
+GGTTAATCGTGTGACCGCGGTGGCTGGCACGAAATTGACCAACCCTGGGGTTAGTATAGCTTAGTTAAACTTTCGTTTATTGCTAAAGGTTAATCACTGCTGTTTCCCGTGGGGGTGTGGCTAGGCTAAGCGTTT
